write json 형식으로 보내는데 이 때 User 타입으로 반환이 안됨 처리 어떻게 해야할 지 모르겠음

+++ sesac/src/main/java/com/example/sesac/user/UserService.java
@@ -1,9 +1,11 @@
 package com.example.sesac.user;
 
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Date;
+import java.sql.SQLIntegrityConstraintViolationException;
+import java.time.LocalDateTime;
 
 @Service
 public class UserService {
@@ -11,23 +13,34 @@ public class UserService {
     private UserDAO dao;
 
     //sign-up
+    @Transactional
     public void saveUser(UserDTO dto){
-        dto.setJoinDate(new Date());
+        dto.setJoinDate(LocalDateTime.now());
         dto.setRole(Role.User);
 
         dao.save(new User(dto.getUid(), dto.getPwd(), dto.getEmail(), dto.getJoinDate(), dto.getRole()));
     }
 
-    //탈퇴
-    public void deleteUser(String uid){
-        dao.deleteByUid(uid);
-    }
-
     //내 정보 조회
     public UserDTO getUser(String uid){
         User user = dao.findByUid(uid);
-        //System.out.println("service에서 받아온 email정보 : " + user.getEmail());
+
         return user == null ? null : new UserDTO( user.getUid(), user.getPwd(),
                 user.getEmail(), user.getJoinDate(), user.getRole() );   //비밀번호 검증 필요
+    }
+
+    // 비밀번호 확인 후 삭제
+    @Transactional
+    public boolean pwdCheckAndDelete(String uid, String check){
+        User user = dao.findByUid(uid);
+
+        // 비밀번호 일치 여부 확인
+        if (user.getPwd().equals(check)) {
+            System.out.println("비밀번호가 일치합니다. 사용자 삭제.");
+            dao.deleteByUid(uid); // 비밀번호가 맞으면 사용자 삭제
+            return true;
+        }
+
+        return false;
     }
 }
